requirements.txt
Flask==1.1.2
itsdangerous==1.1.0
pydot==1.4.1
wincertstore==0.2
gunicorn==20.0.4
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
keras==2.4.3
pillow==8.1.0